Refactor: Chat, WebRTC Logic
