+++ client/src/hooks/useFanUP.ts
@@ -5,6 +5,7 @@ import { initializeMyStream } from '@store/user';
 import { ReducerType } from '@store/rootReducer';
 
 import { socket, SOCKET_EVENTS, connectSocket } from '@/socket';
+import { MyEmail } from '@utils/generateRandomString';
 
 const urls = [
     'stun:stun.l.google.com:19302',
@@ -14,18 +15,6 @@ const urls = [
     'stun:stun4.l.google.com:19302',
 ];
 
-//이메일 연동이 되기전까지, 내 이메일을 랜덤으로 생성하기 위한 것
-const generateRandomString = (num: number) => {
-    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
-    let result = '';
-    const charactersLength = characters.length;
-    for (let i = 0; i < num; i++) {
-        result += characters.charAt(Math.floor(Math.random() * charactersLength));
-    }
-
-    return result;
-};
-
 const useFanUP = (): [
     any[],
     React.MutableRefObject<{
@@ -34,7 +23,7 @@ const useFanUP = (): [
 ] => {
     const [users, setUsers] = useState<any[]>([]);
     const peerConnections = useRef<{ [key: string]: RTCPeerConnection }>({});
-    const myEmail = useMemo(() => generateRandomString(10), []);
+    const myEmail = useMemo(() => MyEmail, []);
 
     const myStream = useSelector<ReducerType, MediaStream | null>(
         ({ userSlice }) => userSlice.myStream
